Get Vagrant working better
Main change in this commit is to add the `vagrant` user to the `admin` group so that
it can sudo without a password. Vagrant expects that to work and various vagrant
commands fail otherwise.

Also, set ALLOWED_HOSTS in the staging config to wildcard since our nginx config
checks the host already and it was 500'ing.

Also, quiet down some of the operations in the fabfile like `pip install` and
`git clone`.

Also, more comments.